refactor(triage): simplify progress output structure
