+++ src/cli/commands/triage/pipeline.test.ts
@@ -6,12 +6,6 @@ import { join } from "node:path";
 import { DEFAULT_PATCHMILL_CONFIG } from "../../../config/defaults.ts";
 import { createTriagePolicy } from "../../../policy/triage.ts";
 import { createStaticCommandRunner } from "../../../../test-support/command-runner.ts";
-import {
-  createTriageProgressReporter,
-  formatProgressIssueLines,
-  formatResultLines,
-  HELP_TEXT,
-} from "./main.ts";
 import { runTriage } from "./pipeline.ts";
 
 const issueJson = JSON.stringify([
@@ -246,191 +240,15 @@ test("runTriage targeted GitHub issue reads issue by number", async () => {
   );
 });
 
-test("HELP_TEXT documents usage and active triage protection wording", () => {
-  assert.match(HELP_TEXT, /Usage:/);
-  assert.match(HELP_TEXT, /--help/);
-  assert.match(HELP_TEXT, /-h/);
-  assert.doesNotMatch(HELP_TEXT, /--execute/);
-  assert.match(HELP_TEXT, /Automated issue triage/);
-  assert.doesNotMatch(HELP_TEXT, /Automated Forgejo issue triage/);
-  assert.match(
-    HELP_TEXT,
-    /Runs the configured triage skill against eligible untriaged open issues by default/,
-  );
-  assert.doesNotMatch(HELP_TEXT, /Defaults to showing this help/);
-  assert.match(
-    HELP_TEXT,
-    /Preview configured triage skill decisions without mutating the configured issue host/,
-  );
-  assert.doesNotMatch(HELP_TEXT, /without mutating Forgejo/);
-  assert.match(HELP_TEXT, /--dry-run/);
-  assert.match(HELP_TEXT, /--issue <number>/);
-  assert.match(HELP_TEXT, /--all/);
-  assert.match(HELP_TEXT, /without active triage or protection labels/);
-  assert.match(
-    HELP_TEXT,
-    /include issues already carrying triage or protection labels such as in-progress or blocked/,
-  );
-});
-
-test("formatProgressIssueLines matches the triage output demo layout", () => {
-  const lines = formatProgressIssueLines(
-    {
-      issueNumber: 879,
-      title: "OAuth callback fails in staging",
-      url: "https://forgejo.example.com/acme/agibase/issues/879",
-      previousLabels: [],
-      finalLabels: ["agent-unsuitable"],
-      primaryBucket: "agent-unsuitable",
-      questions: [],
-      comment: [
-        "> _This was generated by AI during triage._",
-        ">",
-        "> This issue is not suitable for an agent because it needs a product decision",
-        "> before implementation can begin.",
-        "> Please reopen with concrete acceptance criteria if this should be revisited.",
-        "> Hidden sixth line.",
-      ].join("\n"),
-      addedComments: ["comment was added"],
-      previousState: "open",
-      finalState: "closed",
-      mutationStatus: "observed",
-    },
-    2,
-    49,
-  );
-
-  assert.deepEqual(lines, [
-    "──────────────────────────────────────────────────────────────────────────────",
-    "#879 OAuth callback fails in staging",
-    "  link: https://forgejo.example.com/acme/agibase/issues/879",
-    "  labels: (none) -> agent-unsuitable",
-    "  state: open -> closed",
-    "  comment added:",
-    "  > _This was generated by AI during triage._",
-    "  >",
-    "  > This issue is not suitable for an agent because it needs a product decision",
-    "  > before implementation can begin.",
-    "  > Please reopen with concrete acceptance criteria if this should be revisited.",
-    "",
-    "progress: 2/49 triaged",
-    "",
-  ]);
-});
-
-test("createTriageProgressReporter prints header, issue progress, and footer", () => {
-  const output: string[] = [];
-  const reporter = createTriageProgressReporter({
-    command: "patchmill triage --limit 1",
-    writeLine: (line) => output.push(line),
-  });
-
-  reporter.onProgress({ type: "selected", total: 1 });
-  reporter.onProgress({
-    type: "issue",
-    completed: 1,
-    total: 1,
-    issue: {
-      issueNumber: 880,
-      title: "Cache invalidation flakes in worker",
-      url: "https://forgejo.example.com/acme/agibase/issues/880",
-      previousLabels: [],
-      finalLabels: ["agent-ready"],
-      primaryBucket: "agent-ready",
-      questions: [],
-      comment: null,
-      mutationStatus: "observed",
-    },
-  });
-  reporter.finish({
-    status: "applied",
-    issueCount: 1,
-    logPath: ".patchmill/triage-runs/triage.json",
-    issues: [],
-  });
-
-  assert.deepEqual(output, [
-    "> patchmill triage --limit 1",
-    "",
-    "issues: 1",
-    "",
-    "──────────────────────────────────────────────────────────────────────────────",
-    "#880 Cache invalidation flakes in worker",
-    "  link: https://forgejo.example.com/acme/agibase/issues/880",
-    "  labels: (none) -> agent-ready",
-    "",
-    "progress: 1/1 triaged",
-    "",
-    "agent issue triage: applied",
-    "log: .patchmill/triage-runs/triage.json",
-  ]);
-});
-
-test("formatResultLines prints dry-run previews and observed changes", () => {
-  const dryRunLines = formatResultLines({
-    status: "dry-run",
-    issueCount: 1,
-    logPath: "/tmp/triage.json",
-    issues: [
-      {
-        issueNumber: 1,
-        title: "Needs info",
-        previousLabels: ["bug"],
-        finalLabels: ["bug", "needs-info"],
-        primaryBucket: "needs-info",
-        rationale: "Missing reproduction details.",
-        questions: ["What exact steps reproduce the issue?"],
-        comment: "What exact steps reproduce the issue?",
-        mutationStatus: "preview",
-      },
-    ],
-  });
-
-  assert.deepEqual(dryRunLines, [
-    "#1 needs-info preview",
-    "  labels: bug -> bug, needs-info",
-    "  comment: What exact steps reproduce the issue?",
-  ]);
-
-  const executeLines = formatResultLines({
-    status: "applied",
-    issueCount: 1,
-    logPath: "/tmp/triage.json",
-    issues: [
-      {
-        issueNumber: 2,
-        title: "Ready",
-        previousLabels: ["needs-triage"],
-        finalLabels: ["ready-for-agent"],
-        primaryBucket: "agent-ready",
-        questions: [],
-        comment: "## Agent Brief",
-        addedComments: ["## Agent Brief"],
-        previousState: "open",
-        finalState: "open",
-        mutationStatus: "observed",
-      },
-    ],
-  });
-
-  assert.deepEqual(executeLines, [
-    "#2 agent-ready observed",
-    "  labels: needs-triage -> ready-for-agent",
-    "  comment: ## Agent Brief",
-  ]);
-});
-
 test("runTriage executes configured skill by default and reports observed changes", async () => {
   const logDir = await mkdtemp(join(tmpdir(), "triage-pipeline-"));
-  const beforeIssueJson = JSON.stringify([
-    {
-      index: 1,
-      title: "Needs triage",
-      body: "Broken",
-      state: "open",
-      labels: [{ name: "bug" }],
-    },
-  ]);
+  const beforeIssueJson = JSON.stringify({
+    index: 1,
+    title: "Needs triage",
+    body: "Broken",
+    state: "open",
+    labels: [{ name: "bug" }],
+  });
   const afterIssueJson = JSON.stringify({
     index: 1,
     title: "Needs triage",
@@ -440,7 +258,6 @@ test("runTriage executes configured skill by default and reports observed change
   });
   const runner = createStaticCommandRunner([
     { code: 0, stdout: beforeIssueJson, stderr: "" },
-    { code: 0, stdout: JSON.stringify([]), stderr: "" },
     noCommentsOutput,
     { code: 0, stdout: "triaged", stderr: "" },
     { code: 0, stdout: afterIssueJson, stderr: "" },
@@ -737,6 +554,13 @@ test("runTriage execute snapshots Forgejo issues without repeated all-issue scan
 
 test("runTriage passes explicit tea login to Forgejo commands only", async () => {
   const logDir = await mkdtemp(join(tmpdir(), "triage-pipeline-"));
+  const beforeIssueJson = JSON.stringify({
+    index: 1,
+    title: "Needs info",
+    body: "Broken",
+    state: "open",
+    labels: [{ name: "bug" }],
+  });
   const afterIssueJson = JSON.stringify({
     index: 1,
     title: "Needs info",
@@ -745,8 +569,7 @@ test("runTriage passes explicit tea login to Forgejo commands only", async () =>
     labels: [{ name: "agent-ready" }, { name: "bug" }],
   });
   const runner = createStaticCommandRunner([
-    { code: 0, stdout: issueJson, stderr: "" },
-    { code: 0, stdout: JSON.stringify([]), stderr: "" },
+    { code: 0, stdout: beforeIssueJson, stderr: "" },
     noCommentsOutput,
     { code: 0, stdout: "triaged", stderr: "" },
     { code: 0, stdout: afterIssueJson, stderr: "" },
@@ -835,15 +658,13 @@ test("runTriage passes configured custom skills through to the triage agent", as
 
 test("runTriage execute passes configured state map to the Pi prompt", async () => {
   const logDir = await mkdtemp(join(tmpdir(), "triage-pipeline-"));
-  const beforeIssueJson = JSON.stringify([
-    {
-      index: 1,
-      title: "Needs triage",
-      body: "Broken",
-      state: "open",
-      labels: [{ name: "bug" }],
-    },
-  ]);
+  const beforeIssueJson = JSON.stringify({
+    index: 1,
+    title: "Needs triage",
+    body: "Broken",
+    state: "open",
+    labels: [{ name: "bug" }],
+  });
   const afterIssueJson = JSON.stringify({
     index: 1,
     title: "Needs triage",
@@ -858,27 +679,18 @@ test("runTriage execute passes configured state map to the Pi prompt", async () 
 
       if (
         command === "tea" &&
-        args.includes("issues") &&
-        args.includes("list")
-      ) {
-        const page = args[args.indexOf("--page") + 1];
-        if (page) {
-          return {
-            code: 0,
-            stdout: page === "1" ? beforeIssueJson : JSON.stringify([]),
-            stderr: "",
-          };
-        }
-        return { code: 0, stdout: afterIssueJson, stderr: "" };
-      }
-
-      if (
-        command === "tea" &&
         args[0] === "issues" &&
         args[1] === "1" &&
         args.includes("--fields")
       ) {
-        return { code: 0, stdout: afterIssueJson, stderr: "" };
+        const piCalls = runner.calls.filter(
+          (call) => call.command === "pi",
+        ).length;
+        return {
+          code: 0,
+          stdout: piCalls > 0 ? afterIssueJson : beforeIssueJson,
+          stderr: "",
+        };
       }
 
       if (command === "tea" && args.includes("--comments")) {
@@ -934,229 +746,6 @@ test("runTriage execute passes configured state map to the Pi prompt", async () 
 
   assert.equal(result.status, "applied");
   assert.ok(runner.calls.some((call) => call.command === "pi"));
-});
-
-test("runTriage applies limit after listing open issues", async () => {
-  const logDir = await mkdtemp(join(tmpdir(), "triage-pipeline-"));
-  const twoIssues = JSON.stringify([
-    { index: 1, title: "One", body: "", state: "open", labels: [] },
-    { index: 2, title: "Two", body: "", state: "open", labels: [] },
-  ]);
-  const runner = createStaticCommandRunner([
-    { code: 0, stdout: twoIssues, stderr: "" },
-    { code: 0, stdout: JSON.stringify([]), stderr: "" },
-    noCommentsOutput,
-    {
-      code: 0,
-      stdout: previewJson([
-        {
-          issueNumber: 1,
-          currentLabels: [],
-          proposedLabels: ["agent-unsuitable"],
-          canonicalBucket: "agent-unsuitable",
-          rationale: "Not actionable for automation.",
-          wouldComment: null,
-          wouldClose: false,
-          questions: [],
-        },
-      ]),
-      stderr: "",
-    },
-  ]);
-
-  const result = await runTriage(runner, {
-    repoRoot: "/repo",
-    dryRun: true,
-    execute: false,
-    limit: 1,
-    logDir,
-    host: DEFAULT_PATCHMILL_CONFIG.host,
-  });
-
-  assert.equal(result.issueCount, 1);
-});
-
-test("runTriage skips already triaged open issues by default before applying limit", async () => {
-  const logDir = await mkdtemp(join(tmpdir(), "triage-pipeline-"));
-  const mixedIssues = JSON.stringify([
-    {
-      index: 1,
-      title: "Already ready",
-      body: "Clear",
-      state: "open",
-      labels: [{ name: "agent-ready" }],
-    },
-    {
-      index: 2,
-      title: "Already blocked",
-      body: "Needs decision",
-      state: "open",
-      labels: [{ name: "needs-info" }],
-    },
-    {
-      index: 3,
-      title: "Untriaged",
-      body: "Broken",
-      state: "open",
-      labels: [{ name: "bug" }],
-    },
-  ]);
-  const runner = createStaticCommandRunner([
-    { code: 0, stdout: mixedIssues, stderr: "" },
-    { code: 0, stdout: JSON.stringify([]), stderr: "" },
-    noCommentsOutput,
-    { code: 0, stdout: agentReadyPreviewJson(3), stderr: "" },
-  ]);
-
-  const result = await runTriage(runner, {
-    repoRoot: "/repo",
-    dryRun: true,
-    execute: false,
-    limit: 1,
-    logDir,
-    host: DEFAULT_PATCHMILL_CONFIG.host,
-  });
-
-  assert.equal(result.issueCount, 1);
-  assert.equal(result.issues[0]?.issueNumber, 3);
-});
-
-test("runTriage skips in-progress issues by default before applying limit", async () => {
-  const logDir = await mkdtemp(join(tmpdir(), "triage-pipeline-"));
-  const mixedIssues = JSON.stringify([
-    {
-      index: 1,
-      title: "Claimed",
-      body: "Already being worked",
-      state: "open",
-      labels: [{ name: "in-progress" }],
-    },
-    {
-      index: 2,
-      title: "Untriaged",
-      body: "Broken",
-      state: "open",
-      labels: [{ name: "bug" }],
-    },
-  ]);
-  const runner = createStaticCommandRunner([
-    { code: 0, stdout: mixedIssues, stderr: "" },
-    { code: 0, stdout: JSON.stringify([]), stderr: "" },
-    noCommentsOutput,
-    { code: 0, stdout: agentReadyPreviewJson(2), stderr: "" },
-  ]);
-
-  const result = await runTriage(runner, {
-    repoRoot: "/repo",
-    dryRun: true,
-    execute: false,
-    limit: 1,
-    logDir,
-    host: DEFAULT_PATCHMILL_CONFIG.host,
-  });
-
-  assert.equal(result.issueCount, 1);
-  assert.equal(result.issues[0]?.issueNumber, 2);
-});
-
-test("runTriage skips blocked issues by default before applying limit", async () => {
-  const logDir = await mkdtemp(join(tmpdir(), "triage-pipeline-"));
-  const mixedIssues = JSON.stringify([
-    {
-      index: 1,
-      title: "Blocked",
-      body: "Waiting on dependency",
-      state: "open",
-      labels: [{ name: "blocked" }],
-    },
-    {
-      index: 2,
-      title: "Untriaged",
-      body: "Broken",
-      state: "open",
-      labels: [{ name: "bug" }],
-    },
-  ]);
-  const runner = createStaticCommandRunner([
-    { code: 0, stdout: mixedIssues, stderr: "" },
-    { code: 0, stdout: JSON.stringify([]), stderr: "" },
-    noCommentsOutput,
-    { code: 0, stdout: agentReadyPreviewJson(2), stderr: "" },
-  ]);
-
-  const result = await runTriage(runner, {
-    repoRoot: "/repo",
-    dryRun: true,
-    execute: false,
-    limit: 1,
-    logDir,
-    host: DEFAULT_PATCHMILL_CONFIG.host,
-  });
-
-  assert.equal(result.issueCount, 1);
-  assert.equal(result.issues[0]?.issueNumber, 2);
-});
-
-test("runTriage --all includes in-progress issues for recovery", async () => {
-  const logDir = await mkdtemp(join(tmpdir(), "triage-pipeline-"));
-  const triagedIssue = JSON.stringify([
-    {
-      index: 1,
-      title: "Claimed",
-      body: "Already being worked",
-      state: "open",
-      labels: [{ name: "in-progress" }],
-    },
-  ]);
-  const runner = createStaticCommandRunner([
-    { code: 0, stdout: triagedIssue, stderr: "" },
-    { code: 0, stdout: JSON.stringify([]), stderr: "" },
-    noCommentsOutput,
-    { code: 0, stdout: agentReadyPreviewJson(1), stderr: "" },
-  ]);
-
-  const result = await runTriage(runner, {
-    repoRoot: "/repo",
-    dryRun: true,
-    execute: false,
-    all: true,
-    logDir,
-    host: DEFAULT_PATCHMILL_CONFIG.host,
-  });
-
-  assert.equal(result.issueCount, 1);
-  assert.equal(result.issues[0]?.issueNumber, 1);
-});
-
-test("runTriage targeted issue overrides the in-progress default skip", async () => {
-  const logDir = await mkdtemp(join(tmpdir(), "triage-pipeline-"));
-  const triagedIssue = JSON.stringify([
-    {
-      index: 1,
-      title: "Claimed",
-      body: "Already being worked",
-      state: "open",
-      labels: [{ name: "in-progress" }],
-    },
-  ]);
-  const runner = createStaticCommandRunner([
-    { code: 0, stdout: triagedIssue, stderr: "" },
-    { code: 0, stdout: JSON.stringify([]), stderr: "" },
-    noCommentsOutput,
-    { code: 0, stdout: agentReadyPreviewJson(1), stderr: "" },
-  ]);
-
-  const result = await runTriage(runner, {
-    repoRoot: "/repo",
-    dryRun: true,
-    execute: false,
-    issueNumber: 1,
-    logDir,
-    host: DEFAULT_PATCHMILL_CONFIG.host,
-  });
-
-  assert.equal(result.issueCount, 1);
-  assert.equal(result.issues[0]?.issueNumber, 1);
 });
 
 test("runTriage writes a failure log before rethrowing execute agent errors", async () => {
@@ -1322,11 +911,21 @@ test("runTriage writes a failure log when execute snapshot listing fails", async
   assert.match(log.error, /snapshot exploded/);
 });
 
-test("runTriage rejects targeted issues that are not open or not found and logs the error", async () => {
+test("runTriage rejects targeted issues that are not open and logs the error", async () => {
   const logDir = await mkdtemp(join(tmpdir(), "triage-pipeline-"));
   const runner = createStaticCommandRunner([
-    { code: 0, stdout: issueJson, stderr: "" },
-    { code: 0, stdout: JSON.stringify([]), stderr: "" },
+    {
+      code: 0,
+      stdout: JSON.stringify({
+        index: 123,
+        title: "Closed",
+        body: "Already closed",
+        state: "closed",
+        labels: [],
+      }),
+      stderr: "",
+    },
+    noCommentsOutput,
   ]);
 
   await assert.rejects(
