+++ src/cli/commands/triage/progress-output.test.ts
@@ -1,0 +1,154 @@
+import test from "node:test";
+import assert from "node:assert/strict";
+import {
+  createTriageProgressReporter,
+  formatProgressIssueLines,
+  formatResultLines,
+} from "./progress-output.ts";
+
+test("formatProgressIssueLines matches the triage output demo layout", () => {
+  const lines = formatProgressIssueLines(
+    {
+      issueNumber: 879,
+      title: "OAuth callback fails in staging",
+      url: "https://forgejo.example.com/acme/agibase/issues/879",
+      previousLabels: [],
+      finalLabels: ["agent-unsuitable"],
+      primaryBucket: "agent-unsuitable",
+      questions: [],
+      comment: [
+        "> _This was generated by AI during triage._",
+        ">",
+        "> This issue is not suitable for an agent because it needs a product decision",
+        "> before implementation can begin.",
+        "> Please reopen with concrete acceptance criteria if this should be revisited.",
+        "> Hidden sixth line.",
+      ].join("\n"),
+      addedComments: ["comment was added"],
+      previousState: "open",
+      finalState: "closed",
+      mutationStatus: "observed",
+    },
+    2,
+    49,
+  );
+
+  assert.deepEqual(lines, [
+    "──────────────────────────────────────────────────────────────────────────────",
+    "#879 OAuth callback fails in staging",
+    "  link: https://forgejo.example.com/acme/agibase/issues/879",
+    "  labels: (none) -> agent-unsuitable",
+    "  state: open -> closed",
+    "  comment added:",
+    "  > _This was generated by AI during triage._",
+    "  >",
+    "  > This issue is not suitable for an agent because it needs a product decision",
+    "  > before implementation can begin.",
+    "  > Please reopen with concrete acceptance criteria if this should be revisited.",
+    "",
+    "progress: 2/49 triaged",
+    "",
+  ]);
+});
+
+test("createTriageProgressReporter prints header, issue progress, and footer", () => {
+  const output: string[] = [];
+  const reporter = createTriageProgressReporter({
+    command: "patchmill triage --limit 1",
+    writeLine: (line) => output.push(line),
+  });
+
+  reporter.onProgress({ type: "selected", total: 1 });
+  reporter.onProgress({
+    type: "issue",
+    completed: 1,
+    total: 1,
+    issue: {
+      issueNumber: 880,
+      title: "Cache invalidation flakes in worker",
+      url: "https://forgejo.example.com/acme/agibase/issues/880",
+      previousLabels: [],
+      finalLabels: ["agent-ready"],
+      primaryBucket: "agent-ready",
+      questions: [],
+      comment: null,
+      mutationStatus: "observed",
+    },
+  });
+  reporter.finish({
+    status: "applied",
+    issueCount: 1,
+    logPath: ".patchmill/triage-runs/triage.json",
+    issues: [],
+  });
+
+  assert.deepEqual(output, [
+    "> patchmill triage --limit 1",
+    "",
+    "issues: 1",
+    "",
+    "──────────────────────────────────────────────────────────────────────────────",
+    "#880 Cache invalidation flakes in worker",
+    "  link: https://forgejo.example.com/acme/agibase/issues/880",
+    "  labels: (none) -> agent-ready",
+    "",
+    "progress: 1/1 triaged",
+    "",
+    "agent issue triage: applied",
+    "log: .patchmill/triage-runs/triage.json",
+  ]);
+});
+
+test("formatResultLines prints dry-run previews and observed changes", () => {
+  const dryRunLines = formatResultLines({
+    status: "dry-run",
+    issueCount: 1,
+    logPath: "/tmp/triage.json",
+    issues: [
+      {
+        issueNumber: 1,
+        title: "Needs info",
+        previousLabels: ["bug"],
+        finalLabels: ["bug", "needs-info"],
+        primaryBucket: "needs-info",
+        rationale: "Missing reproduction details.",
+        questions: ["What exact steps reproduce the issue?"],
+        comment: "What exact steps reproduce the issue?",
+        mutationStatus: "preview",
+      },
+    ],
+  });
+
+  assert.deepEqual(dryRunLines, [
+    "#1 needs-info preview",
+    "  labels: bug -> bug, needs-info",
+    "  comment: What exact steps reproduce the issue?",
+  ]);
+
+  const executeLines = formatResultLines({
+    status: "applied",
+    issueCount: 1,
+    logPath: "/tmp/triage.json",
+    issues: [
+      {
+        issueNumber: 2,
+        title: "Ready",
+        previousLabels: ["needs-triage"],
+        finalLabels: ["ready-for-agent"],
+        primaryBucket: "agent-ready",
+        questions: [],
+        comment: "## Agent Brief",
+        addedComments: ["## Agent Brief"],
+        previousState: "open",
+        finalState: "open",
+        mutationStatus: "observed",
+      },
+    ],
+  });
+
+  assert.deepEqual(executeLines, [
+    "#2 agent-ready observed",
+    "  labels: needs-triage -> ready-for-agent",
+    "  comment: ## Agent Brief",
+  ]);
+});
