+++ src/host/types.ts
@@ -37,7 +37,6 @@ export type IssueHostProvider = {
   missingLabelRemediation(label: LabelDefinition): string;
   listOpenIssues(): Promise<IssueSummary[]>;
   viewIssue(issueNumber: number): Promise<IssueSummary>;
-  listIssuesByNumbers(issueNumbers: readonly number[]): Promise<IssueSummary[]>;
   hydrateIssueComments(issues: IssueSummary[]): Promise<IssueSummary[]>;
   listLabels(): Promise<string[]>;
   createLabel(label: LabelDefinition): Promise<void>;
